src/jenkins_mcp_server/requirements.txt
pylint
requests
google-generativeai
Flask
python-jenkins
Flask-Limiter
cachetools
pydantic
tenacity
